[refactor] : update project structure
