+++ server/Makefile
@@ -1,5 +1,5 @@
 dev:
-	export APP_ENV=dev && go run ./main.go
+	export APP_ENV=dev && go run ./cmd/myapp/main.go
 
 prod:
-	export APP_ENV=prod && go run ./main.go
+	export APP_ENV=prod && go run ./cmd/myapp/main.go
